+++ packages/messaging/src/testing/fakes/token-details.ts
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-import { TokenDetails } from '../../interfaces/token-details';
 import { FakePushSubscription } from './service-worker';
+import { TokenDetails } from '../../interfaces/token-details';
 import { arrayToBase64 } from '../../helpers/array-base64-translator';
 
 export function getFakeTokenDetails(): TokenDetails {